Feature - update myApps page (vue) (#1371)
* First commit

* Remove commented code

* Add missing css

* Fix css

* Fix form-group cols

* Fix logged in status

* Fix grammar

* Set app fetch order properly
